Move addZoweSession functions to utils and fix tests
