+++ __tests__/__integration__/extension.integration.test.ts
@@ -110,7 +110,7 @@ describe("Extension Integration Tests", () => {
             const stubresolve = sandbox.stub(utils, "resolveQuickPickHelper");
             stubresolve.returns(new utils.FilterItem(profileNamesList[0]));
 
-            await extension.addZoweSession(testTree);
+            await utils.addZoweSession(testTree);
             expect(testTree.mSessionNodes[testTree.mSessionNodes.length - 1].label).to.equal(profileNamesList[0]);
         }).timeout(TIMEOUT);
     });
@@ -876,7 +876,7 @@ describe("Extension Integration Tests - USS", () => {
             const stubresolve = sandbox.stub(utils, "resolveQuickPickHelper");
             stubresolve.returns(new utils.FilterItem(profileNamesList[0]));
 
-            await extension.addZoweSession(ussTestTree);
+            await utils.addZoweSession(ussTestTree);
             expect(ussTestTree.mSessionNodes[ussTestTree.mSessionNodes.length - 1].label).to.equal(profileNamesList[0]);
         }).timeout(TIMEOUT);
     });
